# `pipe` CLI shall respect `-e` and `-f` combination