--- a/pipe-cli/src/utilities/storage/azure.py
+++ b/pipe-cli/src/utilities/storage/azure.py
@@ -214,9 +214,6 @@ class TransferBetweenAzureBucketsManager(AzureManager, AbstractTransferManager):
     def get_source_key(self, source_wrapper, source_path):
         return source_path
 
-    def get_source_size(self, source_wrapper, source_key, source_size):
-        return source_wrapper.get_list_manager().get_file_size(source_key)
-
     def transfer(self, source_wrapper, destination_wrapper, path=None, relative_path=None, clean=False,
                  quiet=False, size=None, tags=(), lock=None):
         full_path = path
@@ -274,9 +271,6 @@ class AzureDownloadManager(AzureManager, AbstractTransferManager):
     def get_source_key(self, source_wrapper, source_path):
         return source_path or source_wrapper.path
 
-    def get_source_size(self, source_wrapper, source_key, source_size):
-        return source_wrapper.get_list_manager().get_file_size(source_key)
-
     def transfer(self, source_wrapper, destination_wrapper, path=None,
                  relative_path=None, clean=False, quiet=False, size=None, tags=None, lock=None):
         source_key = self.get_source_key(source_wrapper, path)
@@ -303,9 +297,6 @@ class AzureUploadManager(AzureManager, AbstractTransferManager):
             return os.path.join(source_wrapper.path, source_path)
         else:
             return source_wrapper.path
-
-    def get_source_size(self, source_wrapper, source_key, source_size):
-        return StorageOperations.get_local_file_size(source_key)
 
     def transfer(self, source_wrapper, destination_wrapper, path=None, relative_path=None, clean=False, quiet=False,
                  size=None, tags=(), lock=None):
@@ -346,9 +337,6 @@ class TransferFromHttpOrFtpToAzureManager(AzureManager, AbstractTransferManager)
 
     def get_source_key(self, source_wrapper, source_path):
         return source_path or source_wrapper.path
-
-    def get_source_size(self, source_wrapper, source_key, source_size):
-        return source_size
 
     def transfer(self, source_wrapper, destination_wrapper, path=None, relative_path=None, clean=False, quiet=False,
                  size=None, tags=(), lock=None):