… pipe-cli/src/utilities/storage/gs.py
… pipe-cli/src/utilities/storage/local.py
… pipe-cli/src/utilities/storage/s3.py