The current PR provides fixes introduces in issue #2039 

`-f` flag is not required for copy into existing folder. If the same file exists into source and destination (and shall be copied according to `-e`/`-i`/`-s` options) an error will be occurred in case `-f` is not specified. 

WARN:
The performance may decrease if `-f` flag is not specified since we shall check all destinations