@NShaforostov 2021-06-30
I've tested this new functionality in `AWS` and `Azure` deployments.
Also, the previous functionality was verified.
In general, all works well.

Minor note was encountered in `Azure` deployment:
when performing `cp`/`mv` command from disk to storage or between two storages, the commands are being performed fully and at the end (after some pause) the message appears - "Error: API respond with http status 404."
It looks in the following way:
![image](https://user-images.githubusercontent.com/45459424/123965987-aa0da780-d9bd-11eb-8995-58dee6136236.png)

As for the performance, it really decreases on operations with lots of files. See benchmarking (I've been coping/moving different count of files from disk to storage):
|  | Previously |  | Current |  | _Delta_ |
|---|---|---|---|---|---|
|  |  | With **`-f`** |  | With **`-f`** |  |
| 15 files, each 10MB, on 5 hierarchy nesting levels | 0m 7.33s | 0m 8.8s | 0m 8.53s | 0m 6.82s | 1.21s |
| 15 files, each 100MB, on 5 hierarchy nesting levels | 0m 17.91s | 0m 18.69s | 0m 18.77s | 0m 18.69s | 0.86s |
| 150 files, each 10MB, on 10 hierarchy nesting levels | 0m 58.32s | 1m 5.8s | 1m 7.51s | 1m 4.86s | 9.18s |
| 180 files, each 50MB, on 30 hierarchy nesting levels | 2m 12.29s | 2m 27.57s | 2m 27.8s | 2m 13.01s | 15.51s |
| 1500 files, each 1MB, on 10 hierarchy nesting levels | 3m 54.01s | 3m 47.54s | 4m 53.44s | 3m 53.89s | 59.42s |
| 1500 files, each 1MB, on 100 hierarchy nesting levels | 4m 1.89s | 3m 59.8s | 5m 18.23s | 4m 16.1s | 1m 16.34s |
| 1500 files, each 5MB, on 20 hierarchy nesting levels | 5m 10.4s | 5m 51.99s | 6m 23.35s | 5m 23.76s | 1m 12.95s |
| 10500 files, each 1MB, on 1 hierarchy nesting level | 26m 50.17s |  | 33m 44.53s |  | 6m 54.36s |

Where:
**Previously** column - column with `cp`/`mv` command performing duration via the "old" `pipe` version
**Current** column - column with `cp`/`mv` command performing duration via the current `pipe` implementation from the PR
**_Delta_** column - column with delta between **Current** and **Previously** values