@NShaforostov 2021-07-02
> * Default behavior of the `pipe storage cp` shall be left untouched (required `-f` and do not analyse the destination contents)
> * The new behavior shall be enabled explicitly via `--verify-destination|-vd` flag
> 
> So the advanced users, who are aware of such capabilities may enable it for their needs. But the general users won't be affected.

I've tested new functionality.
Works well